Corrected grammar in explanation
I changed the comment to explain the working of the loop from past tense to present tense. 

You can see other comments which are in present tense "// runs if condition is false" but comments in other loops are in past tense "//Condition was true. Code in this block will run.". 

Minor but completes the doc.